requirements.txt
asgiref==3.5.0
astroid==2.9.3
colorama==0.4.4
distlib==0.3.4
Django==4.0.3
django-jazzmin==2.4.9
django-phonenumber-field==6.1.0
filelock==3.6.0
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.6.1
phonenumbers==8.12.44
phonenumberslite==8.12.44
Pillow==9.0.1
platformdirs==2.5.1
pylint==2.12.2
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tzdata==2021.5
virtualenv==20.13.3
wrapt==1.13.3
